♻ 프로필 팔로우 버튼 컴포넌트 분리
