+++ src/components/Profile/FollowButton/styles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const Button = styled.button`
+  margin-left: auto;
+  border-radius: 5px;
+  border: 1px solid #ccc;
+  padding: 2px 5px;
+`;
